add new layout for admin panel

+++ app/views/web/UserController/partials/_users.cfm
@@ -4,17 +4,28 @@
         // Determine the status text based on the value of users.status[i]
         var statusText = (users.status[i] == 1) ? "Active" : "Inactive";
 
-        writeOutput('<tr> <td>' & users.id[i] & '</td>');
+        writeOutput('<tr> <td class="text-center">' & i & '</td>');
         writeOutput('<td>' & users.firstname[i] & '</td>');
         writeOutput('<td>' & users.lastname[i] & '</td>');
         writeOutput('<td>' & users.email[i] & '</td>');
         writeOutput('<td>' & statusText & '</td>'); // Show "Active" or "Inactive"
         writeOutput('<td>' & users.name[i] & '</td>');
 
-        writeOutput('<td>' & 
-            '<a href="user/addUser?id=#users.id[i]#">Edit</a> | 
-            <a href="user/delete?id=#users.id[i]#" 
-            onclick="return confirmDelete(#users.id[i]#);">Delete</a>' & '</td></tr>');
+        writeOutput('<td class="text-center">
+            <div class="dropdown">
+                <div class="fw-bold cursor-pointer" data-bs-toggle="dropdown" aria-expanded="false">
+                    ...
+                </div>
+                <ul class="dropdown-menu">
+                    <li>
+                        <a href="user/add?id=#users.id[i]#" class="dropdown-item text-success fs-16">Edit</a>
+                    </li>
+                    <li>
+                        <a href="user/delete?id=#users.id[i]#" class="dropdown-item text-danger fs-16" onclick="return confirmDelete(#users.id[i]#);">Delete</a>
+                    </li>
+                </ul>
+            </div>
+        </td>');
     }
 </cfscript>
 
@@ -28,4 +39,14 @@
         }
         return false; // Prevent the default action (following the link)
     }
+</script>
+<script>
+    let table = new DataTable('#userTable', {
+        columnDefs: [
+            {
+                targets: [5,6], // Adjust these indexes based on your actual table structure
+                orderable: false
+            }
+        ]
+    });
 </script>
